Fix imports, EM function, some text edits
The notebook did not run as published, so I fixed it up so anyone who
downloads this version will be able to immediately run it without running
into errors.

I put the EM algorithm in a function called run so you can easily repeat
the algorithm with new starting values over and over by pressing Ctrl-Enter.

I slightly modified the first paragraph describing the experiment and hope
that the goal is clearer to new readers.